+++ subscription-service/src/main/java/com/fis/evangelist/subscription/VO/Book.java
@@ -8,6 +8,6 @@ public class Book {
 	private String bookId;
 	private String name;
 	private String author;
-	private Long copiesAvailable;
-	private Long totalCopies;
+	private int copiesAvailable;
+	private int totalCopies;
 }
